Revise blending algorithm and implement translucent frame blending.
Additionally:

* Added more timing stats to the driver module.

* Updated README giving more precise description on how to configure and run demo in "diagnosis" mode to find accurate LED reset timing.

* Refactored `ColorDotTarget` into `ComputedColorDotTarget`
and `BlendedColorDotTarget`:

- The former is for technical demonstration of real-time floating
point pixel generation;

- The latter is for demonstrating more efficient pre-computed
pixels rendered via alpha blending.

* Removed pre-multiplied alpha blending, because the results
seem to be problematic:

Rounding errors (off-by-one) are inevitable because we are
combining two separately computed (and rounded) pixels.

On regular display this is not a problem because the pixel
response is pretty linear: 0x01 looks not much different from
0x00, or 0x02.

But LED strips displays 0x00, 0x01 and 0x02 with vastly different
luminosities, so a small rounding error becomes very visible
defects. :(

Looks like we are stuck with regular blending for now.

I may explore 10-bit pre-multiplied color components and see if
that helps, but even if it does, I suspect it won't be computationally
more economical than regular blending...